Fixed initializer of tickbox

+++ src/UI/UIUtilies.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 
 struct screenPos{
-    uint32_t  x;
-    uint32_t  y;
+    uint8_t  x;
+    uint8_t  y;
 };
